refactor: move home views into sites (#1598)
