+++ sites/brc-analytics/views/HomeView/homeView.tsx
